app/requirements.txt
flet
python-dotenv
google-generativeai
uvicorn
fastapi